requirements.txt
blinker==1.9.0
click==8.2.0
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3
flask-cors==6.0.0
flask-migrate==4.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
Flask-Limiter==3.5.0
bleach==6.1.0